server-ts: Convert routes/api/sender
