+++ src/routes/routes.js
@@ -44,7 +44,7 @@ const databaseRoute = require('./api/database');
 const imageRoute = require('./api/image');
 const attributesRoute = require('./api/attributes');
 const scriptRoute = require('./api/script');
-const senderRoute = require('./api/sender.js');
+const senderRoute = require('./api/sender');
 const filesRoute = require('./api/files');
 const searchRoute = require('./api/search');
 const bulkActionRoute = require('./api/bulk_action');
